requirements.txt
requests
bs4
gspread
python-dotenv